refactor(frontend): decompose staff course and assignment pages
Why: issue #14 tracks oversized route files that are hard to maintain and risky to change.

What: extracted staff course roster/modules tabs into dedicated components and moved student assignment submission rendering into a standalone SubmissionCard component, leaving route files more orchestration-focused.

Tests: ran cd frontend; npm run build and cd frontend; npm run lint (warnings only, no errors).

Risk: Medium; behavior is preserved but moved across files, so regressions would likely appear in staff roster/modules interactions or submission test-result expansion UI.
