+++ frontend/src/app/staff/courses/[id]/page.tsx
@@ -1,13 +1,11 @@
 "use client";
 
-import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { useParams, useRouter } from "next/navigation";
 import Link from "next/link";
 import { motion, AnimatePresence } from "framer-motion";
 import {
   ArrowLeft,
-  ArrowUp,
-  ArrowDown,
   BookOpen,
   Users,
   FileText,
@@ -17,33 +15,24 @@ import {
   Upload,
   Loader2,
   AlertCircle,
-  Calendar,
   Clock,
   ChevronRight,
   ChevronLeft,
   ArrowRight,
+  ExternalLink,
   Filter,
   RefreshCw,
-  MoreVertical,
   Pencil,
   Save,
   Trash2,
   Copy,
   UserX,
-  Link as LinkIcon,
-  Download,
-  ExternalLink,
   Beaker,
-  Eye,
-  EyeOff,
-  GripVertical,
   X,
   ChevronDown,
-  ChevronUp,
 } from "lucide-react";
 import {
   courseStaff,
-  staffCourseGitHub,
   staffSubmissions,
   type StaffSubmissionQueueItem,
   type MissingSubmissionsSummaryItem,
@@ -52,28 +41,18 @@ import {
   type Module,
   type Assignment,
   type CourseMembership,
-  type CourseGitHubClaim,
-  type GitHubClassroom,
-  type GitHubClassroomAssignment,
-  type GitHubRosterSync,
-  type ModuleResource,
-  type OrgMembership,
-  type CourseMembershipCreate,
-  type CourseMembershipUpdate,
   type CourseUpdate,
   type LatePolicy,
-  type ImportCsvIssue,
-  type ImportCsvResult,
   ApiError,
 } from "@/lib/api";
 import { useAuthStore, getCourseRole } from "@/lib/store";
 import { PageHeader } from "@/components/shared/PageHeader";
 import { EmptyState } from "@/components/shared/EmptyState";
 import { DataList } from "@/components/shared/DataList";
-import { PdfPreviewModal } from "@/components/shared/PdfPreviewModal";
 import { ConfirmModal } from "@/components/ui/Modal";
 import { reportError } from "@/lib/reportError";
-import { PROGRAMMES, type Programme } from "@/lib/programmes";
+import { RosterTab } from "@/components/staff/course/RosterTab";
+import { ModulesTab } from "@/components/staff/course/ModulesTab";
 
 const fadeInUp = {
   hidden: { opacity: 0, y: 20 },
@@ -88,49 +67,10 @@ const staggerContainer = {
   },
 };
 
-const IMPORT_ISSUE_REASON_LABELS: Record<string, string> = {
-  invalid_email: "Invalid email format",
-  missing_name: "Missing student name",
-  missing_student_number: "Missing student number",
-  missing_programme: "Missing programme",
-  duplicate_student_number_in_csv: "Duplicate student number in CSV",
-  student_number_taken_in_course: "Student number already exists in this course",
-};
-
-const REQUIRED_ROSTER_HEADERS = ["email", "name", "student_number", "programme"] as const;
-
-type CsvPreviewIssue = {
-  rowNumber: number | null;
-  reasonCode: string;
-  reasonLabel: string;
-  email: string;
-  fullName: string;
-  studentNumber: string;
-  programme: string;
-};
-
-type CsvPreviewRow = {
-  rowNumber: number;
-  email: string;
-  fullName: string;
-  studentNumber: string;
-  programme: string;
-  isValid: boolean;
-};
-
-type CsvPreviewData = {
-  fileName: string;
-  headerRow: string[];
-  missingHeaders: string[];
-  rows: CsvPreviewRow[];
-  issues: CsvPreviewIssue[];
-};
-
 type TabType = "overview" | "submissions" | "roster" | "assignments" | "modules";
 
 export default function StaffCoursePage() {
   const params = useParams();
-  const router = useRouter();
   const courseId = Number(params.id);
   const { user } = useAuthStore();
 
@@ -1233,1579 +1173,6 @@ function CourseSubmissionsTab({ courseId }: CourseSubmissionsTabProps) {
 }
 
 // Roster Tab
-interface RosterTabProps {
-  course: Course;
-  memberships: CourseMembership[];
-  onRefresh: () => Promise<void>;
-  focusImportCsvOnMount?: boolean;
-  onConsumedFocusImportCsv?: () => void;
-}
-
-function RosterTab({
-  course,
-  memberships,
-  onRefresh,
-  focusImportCsvOnMount = false,
-  onConsumedFocusImportCsv,
-}: RosterTabProps) {
-  const [orgMembers, setOrgMembers] = useState<OrgMembership[]>([]);
-  const [memberSearch, setMemberSearch] = useState("");
-  const [selectedUserId, setSelectedUserId] = useState<number | null>(null);    
-  const [newRole, setNewRole] = useState<"co_lecturer" | "ta">("ta");
-  const [isAddingStaff, setIsAddingStaff] = useState(false);
-  const [isAddingStudent, setIsAddingStudent] = useState(false);
-  const [isLoadingOrg, setIsLoadingOrg] = useState(false);
-  const [noticeArea, setNoticeArea] = useState<"student" | "staff" | null>(null);
-  const [error, setError] = useState("");
-  const [success, setSuccess] = useState<string>("");
-  const [showStaffSection, setShowStaffSection] = useState(false);
-  const [inviteLinks, setInviteLinks] = useState<string[]>([]);
-  const [lastImportResult, setLastImportResult] = useState<ImportCsvResult | null>(null);
-  const [selectedCsvFile, setSelectedCsvFile] = useState<File | null>(null);
-  const [csvPreview, setCsvPreview] = useState<CsvPreviewData | null>(null);
-  const [isBuildingCsvPreview, setIsBuildingCsvPreview] = useState(false);
-  const [notifyNewSubmissions, setNotifyNewSubmissions] = useState(true);
-  const [isSavingNotifyNewSubmissions, setIsSavingNotifyNewSubmissions] = useState(false);
-  const [notifyPrefError, setNotifyPrefError] = useState("");
-  const [confirmRemoveMembershipId, setConfirmRemoveMembershipId] = useState<number | null>(null);
-  const [isRemovingMember, setIsRemovingMember] = useState(false);
-
-  const [pendingGitHubClaims, setPendingGitHubClaims] = useState<CourseGitHubClaim[]>([]);
-  const [isLoadingGitHubClaims, setIsLoadingGitHubClaims] = useState(false);
-  const [gitHubClaimsError, setGitHubClaimsError] = useState("");
-  const [claimActionId, setClaimActionId] = useState<number | null>(null);
-  const [classrooms, setClassrooms] = useState<GitHubClassroom[]>([]);
-  const [isLoadingClassrooms, setIsLoadingClassrooms] = useState(false);
-  const [githubClassroomsError, setGithubClassroomsError] = useState("");
-  const [selectedClassroomId, setSelectedClassroomId] = useState<number | null>(
-    course.github_classroom_id ?? null
-  );
-  const [isSavingClassroomBinding, setIsSavingClassroomBinding] = useState(false);
-  const [selectedGitHubAssignmentId, setSelectedGitHubAssignmentId] = useState<number | null>(null);
-  const [isRunningRosterSync, setIsRunningRosterSync] = useState(false);
-  const [rosterSync, setRosterSync] = useState<GitHubRosterSync | null>(null);
-  const [githubAssignments, setGithubAssignments] = useState<GitHubClassroomAssignment[]>([]);
-
-  const [studentEmail, setStudentEmail] = useState("");
-  const [studentName, setStudentName] = useState("");
-  const [studentNumber, setStudentNumber] = useState("");
-  const [studentProgramme, setStudentProgramme] = useState<Programme | "">("");
-
-  const enrollStudentsRef = useRef<HTMLDivElement>(null);
-  const importCsvButtonRef = useRef<HTMLButtonElement>(null);
-
-  useEffect(() => {
-    async function loadOrgMembers() {
-      setIsLoadingOrg(true);
-      try {
-        const data = await courseStaff.listOrgMembers(course.id);
-        setOrgMembers(data);
-      } catch (err) {
-        reportError("Failed to load org members", err);
-      } finally {
-        setIsLoadingOrg(false);
-      }
-    }
-    loadOrgMembers();
-  }, [course.id]);
-
-  useEffect(() => {
-    setSelectedClassroomId(course.github_classroom_id ?? null);
-  }, [course.github_classroom_id]);
-
-  const refreshGitHubClaims = useCallback(async () => {
-    setIsLoadingGitHubClaims(true);
-    setGitHubClaimsError("");
-    try {
-      const claims = await staffCourseGitHub.listClaims(course.id, "pending");
-      setPendingGitHubClaims(claims);
-    } catch (err) {
-      if (err instanceof ApiError) setGitHubClaimsError(err.detail);
-      else setGitHubClaimsError("Failed to load GitHub linking requests");
-    } finally {
-      setIsLoadingGitHubClaims(false);
-    }
-  }, [course.id]);
-
-  useEffect(() => {
-    void refreshGitHubClaims();
-  }, [refreshGitHubClaims]);
-
-  const refreshGitHubClassrooms = useCallback(async () => {
-    setIsLoadingClassrooms(true);
-    setGithubClassroomsError("");
-    try {
-      const data = await staffCourseGitHub.listClassrooms(course.id);
-      setClassrooms(data);
-    } catch (err) {
-      if (err instanceof ApiError) setGithubClassroomsError(err.detail);
-      else setGithubClassroomsError("Failed to load GitHub classrooms");
-    } finally {
-      setIsLoadingClassrooms(false);
-    }
-  }, [course.id]);
-
-  useEffect(() => {
-    void refreshGitHubClassrooms();
-  }, [refreshGitHubClassrooms]);
-
-  useEffect(() => {
-    if (!focusImportCsvOnMount) return;
-    // Browsers may block opening a file picker without a direct user gesture,
-    // so we scroll + focus the Import CSV button instead.
-    enrollStudentsRef.current?.scrollIntoView({ behavior: "smooth", block: "start" });
-    importCsvButtonRef.current?.focus();
-    onConsumedFocusImportCsv?.();
-  }, [focusImportCsvOnMount, onConsumedFocusImportCsv]);
-
-  useEffect(() => {
-    let cancelled = false;
-    async function loadNotificationPrefs() {
-      try {
-        const prefs = await courseStaff.getNotificationPreferences(course.id);
-        if (cancelled) return;
-        setNotifyNewSubmissions(Boolean(prefs.notify_new_submissions));
-      } catch {
-        // Default to enabled if prefs aren't available.
-      }
-    }
-    loadNotificationPrefs();
-    return () => {
-      cancelled = true;
-    };
-  }, [course.id]);
-
-  const students = memberships.filter((m) => m.role === "student");
-  const staffMembers = memberships.filter((m) => m.role !== "student");
-
-  const pendingClaimByMembershipId = useMemo(() => {
-    return new Map(pendingGitHubClaims.map((c) => [c.course_membership_id, c]));
-  }, [pendingGitHubClaims]);
-
-  const availableOrgMembers = useMemo(() => {
-    const assigned = new Set(memberships.map((m) => m.user_id));
-    const q = memberSearch.trim().toLowerCase();
-    const filtered = orgMembers.filter((m) => {
-      if (assigned.has(m.user_id)) return false;
-      if (!q) return true;
-      return (m.user_email ?? "").toLowerCase().includes(q);
-    });
-    return filtered.sort((a, b) => (a.user_email ?? "").localeCompare(b.user_email ?? ""));
-  }, [orgMembers, memberships, memberSearch]);
-
-  const resolvedInviteLinks = useMemo(() => {
-    if (typeof window === "undefined") return inviteLinks;
-    const origin = window.location.origin;
-    return inviteLinks.map((link) => {
-      if (link.startsWith("http://") || link.startsWith("https://")) return link;
-      if (link.startsWith("/")) return `${origin}${link}`;
-      return `${origin}/${link}`;
-    });
-  }, [inviteLinks]);
-
-  async function approveGitHubClaim(claim: CourseGitHubClaim) {
-    setClaimActionId(claim.id);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    try {
-      await staffCourseGitHub.approveClaim(course.id, claim.id);
-      await onRefresh();
-      await refreshGitHubClaims();
-      setSuccess(`Linked GitHub account @${claim.github_login}.`);
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to approve GitHub linking");
-    } finally {
-      setClaimActionId(null);
-    }
-  }
-
-  async function rejectGitHubClaim(claim: CourseGitHubClaim) {
-    setClaimActionId(claim.id);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    try {
-      await staffCourseGitHub.rejectClaim(course.id, claim.id);
-      await refreshGitHubClaims();
-      setSuccess(`Rejected GitHub linking request from @${claim.github_login}.`);
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to reject GitHub linking");
-    } finally {
-      setClaimActionId(null);
-    }
-  }
-
-  async function saveGitHubClassroomBinding() {
-    setIsSavingClassroomBinding(true);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    try {
-      const selected = selectedClassroomId
-        ? classrooms.find((item) => item.id === selectedClassroomId) ?? null
-        : null;
-      await courseStaff.updateCourse(course.id, {
-        github_classroom_id: selected?.id ?? null,
-        github_classroom_name: selected?.name ?? null,
-      });
-      await onRefresh();
-      setSuccess(selected ? `Bound to GitHub Classroom: ${selected.name}.` : "GitHub Classroom binding cleared.");
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to save classroom binding");
-    } finally {
-      setIsSavingClassroomBinding(false);
-    }
-  }
-
-  async function runGitHubRosterSync() {
-    setIsRunningRosterSync(true);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    try {
-      const result = await staffCourseGitHub.getRosterSync(course.id, {
-        classroom_id: selectedClassroomId ?? undefined,
-        assignment_id: selectedGitHubAssignmentId ?? undefined,
-      });
-      setRosterSync(result);
-      setGithubAssignments(result.assignments);
-      setSuccess("GitHub roster sync check updated.");
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to run GitHub roster sync");
-    } finally {
-      setIsRunningRosterSync(false);
-    }
-  }
-
-  useEffect(() => {
-    if (!selectedClassroomId) {
-      setGithubAssignments([]);
-      setSelectedGitHubAssignmentId(null);
-      return;
-    }
-    let cancelled = false;
-    async function loadAssignments() {
-      try {
-        const result = await staffCourseGitHub.getRosterSync(course.id, {
-          classroom_id: selectedClassroomId ?? undefined,
-        });
-        if (cancelled) return;
-        setGithubAssignments(result.assignments);
-        setRosterSync(result);
-      } catch {
-        // Assignment options are optional pre-check data.
-      }
-    }
-    void loadAssignments();
-    return () => {
-      cancelled = true;
-    };
-  }, [course.id, selectedClassroomId]);
-
-  async function toggleNotifyNewSubmissions() {
-    const next = !notifyNewSubmissions;
-    setIsSavingNotifyNewSubmissions(true);
-    setNotifyPrefError("");
-    try {
-      const updated = await courseStaff.setNotificationPreferences(course.id, {
-        notify_new_submissions: next,
-      });
-      setNotifyNewSubmissions(Boolean(updated.notify_new_submissions));
-    } catch (err) {
-      if (err instanceof ApiError) setNotifyPrefError(err.detail);
-      else setNotifyPrefError("Failed to update notification preferences");
-    } finally {
-      setIsSavingNotifyNewSubmissions(false);
-    }
-  }
-
-  async function addStaffMember() {
-    if (selectedUserId === null) return;
-    setIsAddingStaff(true);
-    setNoticeArea("staff");
-    setError("");
-    setSuccess("");
-    setInviteLinks([]);
-    try {
-      const payload: CourseMembershipCreate = { user_id: selectedUserId, role: newRole };
-      await courseStaff.enrollUser(course.id, payload);
-      await onRefresh();
-      setSelectedUserId(null);
-      setMemberSearch("");
-      setSuccess("Staff member added.");
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to add staff member");
-    } finally {
-      setIsAddingStaff(false);
-    }
-  }
-
-  async function updateRole(membershipId: number, role: CourseMembership["role"]) {
-    setNoticeArea("staff");
-    setError("");
-    try {
-      const payload: CourseMembershipUpdate = { role };
-      await courseStaff.updateMembership(course.id, membershipId, payload);
-      await onRefresh();
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to update role");
-    }
-  }
-
-  function requestRemoveMember(membershipId: number) {
-    setConfirmRemoveMembershipId(membershipId);
-  }
-
-  async function removeMemberConfirmed() {
-    if (confirmRemoveMembershipId === null) return;
-    setIsRemovingMember(true);
-    setNoticeArea("staff");
-    setError("");
-    try {
-      await courseStaff.removeMembership(course.id, confirmRemoveMembershipId);
-      await onRefresh();
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to remove member");
-    } finally {
-      setIsRemovingMember(false);
-      setConfirmRemoveMembershipId(null);
-    }
-  }
-
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  async function addStudentByEmail() {
-    const email = studentEmail.trim();
-    const full_name = studentName.trim();
-    const student_number = studentNumber.trim();
-    const programme = studentProgramme.trim();
-    if (!email || !full_name || !student_number || !programme) return;
-
-    setIsAddingStudent(true);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    setInviteLinks([]);
-    try {
-      const res = await courseStaff.inviteStudentByEmail(course.id, {
-        email,
-        full_name,
-        student_number,
-        programme,
-      });
-
-      await onRefresh();
-
-      if (res.auto_enrolled > 0) {
-        setSuccess("Student enrolled (existing account).");
-      } else if (res.created_invites > 0) {
-        setSuccess(
-          "Invite created (no email yet). Copy the invite link below and share it manually."
-        );
-        setInviteLinks(res.invite_links ?? []);
-      } else {
-        setSuccess("No invite created.");
-      }
-
-      setStudentEmail("");
-      setStudentName("");
-      setStudentNumber("");
-      setStudentProgramme("");
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to invite student");
-    } finally {
-      setIsAddingStudent(false);
-    }
-  }
-
-  function parseCsvLine(line: string): string[] {
-    const values: string[] = [];
-    let current = "";
-    let inQuotes = false;
-
-    for (let i = 0; i < line.length; i += 1) {
-      const char = line[i];
-      if (char === '"') {
-        const next = line[i + 1];
-        if (inQuotes && next === '"') {
-          current += '"';
-          i += 1;
-          continue;
-        }
-        inQuotes = !inQuotes;
-        continue;
-      }
-      if (char === "," && !inQuotes) {
-        values.push(current);
-        current = "";
-        continue;
-      }
-      current += char;
-    }
-    values.push(current);
-    return values;
-  }
-
-  async function buildCsvPreview(file: File): Promise<CsvPreviewData> {
-    const text = await file.text();
-    const lines = text.replace(/\r\n/g, "\n").replace(/\r/g, "\n").split("\n");
-
-    const headerCells = parseCsvLine(lines[0] ?? "").map((value) => value.trim());
-    const headerIndex = new Map<string, number>();
-    headerCells.forEach((header, index) => {
-      if (!header) return;
-      const key = header.toLowerCase();
-      if (!headerIndex.has(key)) {
-        headerIndex.set(key, index);
-      }
-    });
-
-    const missingHeaders = REQUIRED_ROSTER_HEADERS.filter((header) => !headerIndex.has(header));
-    const issues: CsvPreviewIssue[] = missingHeaders.map((header) => ({
-      rowNumber: null,
-      reasonCode: "missing_required_headers",
-      reasonLabel: `Missing required header: ${header}`,
-      email: "",
-      fullName: "",
-      studentNumber: "",
-      programme: "",
-    }));
-
-    const rows: CsvPreviewRow[] = [];
-    const seenStudentNumbers = new Set<string>();
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-
-    const getCellValue = (cells: string[], key: string): string => {
-      const index = headerIndex.get(key);
-      if (index === undefined) return "";
-      return (cells[index] ?? "").trim();
-    };
-
-    for (let index = 1; index < lines.length; index += 1) {
-      const line = lines[index];
-      if (!line || !line.trim()) continue;
-
-      const rowNumber = index + 1;
-      const cells = parseCsvLine(line);
-      const email = getCellValue(cells, "email");
-      const fullName = getCellValue(cells, "name");
-      const studentNumber = getCellValue(cells, "student_number");
-      const programme = getCellValue(cells, "programme");
-
-      if (!email && !fullName && !studentNumber && !programme) continue;
-
-      const rowIssues: CsvPreviewIssue[] = [];
-      const pushRowIssue = (reasonCode: string, reasonLabel: string) => {
-        rowIssues.push({
-          rowNumber,
-          reasonCode,
-          reasonLabel,
-          email,
-          fullName,
-          studentNumber,
-          programme,
-        });
-      };
-
-      if (!email || !emailRegex.test(email)) {
-        pushRowIssue("invalid_email", IMPORT_ISSUE_REASON_LABELS.invalid_email);
-      }
-      if (!fullName) {
-        pushRowIssue("missing_name", IMPORT_ISSUE_REASON_LABELS.missing_name);
-      }
-      if (!studentNumber) {
-        pushRowIssue("missing_student_number", IMPORT_ISSUE_REASON_LABELS.missing_student_number);
-      }
-      if (!programme) {
-        pushRowIssue("missing_programme", IMPORT_ISSUE_REASON_LABELS.missing_programme);
-      }
-      if (studentNumber) {
-        if (seenStudentNumbers.has(studentNumber)) {
-          pushRowIssue(
-            "duplicate_student_number_in_csv",
-            IMPORT_ISSUE_REASON_LABELS.duplicate_student_number_in_csv
-          );
-        } else {
-          seenStudentNumbers.add(studentNumber);
-        }
-      }
-
-      issues.push(...rowIssues);
-      rows.push({
-        rowNumber,
-        email,
-        fullName,
-        studentNumber,
-        programme,
-        isValid: rowIssues.length === 0 && missingHeaders.length === 0,
-      });
-    }
-
-    return {
-      fileName: file.name,
-      headerRow: headerCells,
-      missingHeaders,
-      rows,
-      issues,
-    };
-  }
-
-  async function handleCsvFileSelected(e: React.ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    e.target.value = "";
-    if (!file) return;
-
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    setInviteLinks([]);
-    setLastImportResult(null);
-    setSelectedCsvFile(file);
-    setIsBuildingCsvPreview(true);
-
-    try {
-      const preview = await buildCsvPreview(file);
-      setCsvPreview(preview);
-
-      const rowIssueCount = preview.issues.filter((issue) => issue.rowNumber !== null).length;
-      if (preview.missingHeaders.length > 0) {
-        setError(
-          `Preview failed: missing headers (${preview.missingHeaders.join(", ")}). Download the template and retry.`
-        );
-      } else {
-        const validRows = preview.rows.filter((row) => row.isValid).length;
-        setSuccess(
-          `Preview ready: ${preview.rows.length} row(s), ${validRows} valid, ${rowIssueCount} row issue(s).`
-        );
-      }
-    } catch {
-      setCsvPreview(null);
-      setSelectedCsvFile(null);
-      setError("Failed to parse CSV preview. Check file encoding/format and try again.");
-    } finally {
-      setIsBuildingCsvPreview(false);
-    }
-  }
-
-  async function importSelectedCsv() {
-    if (!selectedCsvFile || !csvPreview) {
-      setNoticeArea("student");
-      setError("Select and preview a CSV file first.");
-      return;
-    }
-    if (csvPreview.missingHeaders.length > 0) {
-      setNoticeArea("student");
-      setError("Cannot import: CSV is missing required headers.");
-      return;
-    }
-
-    setIsAddingStudent(true);
-    setNoticeArea("student");
-    setError("");
-    setSuccess("");
-    setInviteLinks([]);
-    setLastImportResult(null);
-    try {
-      const res = await courseStaff.importRosterCsv(course.id, selectedCsvFile);
-      setLastImportResult(res);
-      const msg = `Imported: ${res.created_invites} invites created, ${res.auto_enrolled} auto-enrolled.`;
-      if (res.issues.length > 0) {
-        setError(`${msg} Some rows had issues.`);
-      } else {
-        setSuccess(msg);
-      }
-      if ((res.invite_links ?? []).length > 0) setInviteLinks(res.invite_links);
-      await onRefresh();
-      setSelectedCsvFile(null);
-      setCsvPreview(null);
-    } catch (err) {
-      setLastImportResult(null);
-      if (err instanceof ApiError) setError(err.detail);
-      else setError("Failed to import CSV");
-    } finally {
-      setIsAddingStudent(false);
-    }
-  }
-
-  async function copyInviteLinksAll() {
-    setNoticeArea("student");
-    try {
-      await navigator.clipboard.writeText(resolvedInviteLinks.join("\n"));
-      setSuccess("Invite links copied to clipboard.");
-    } catch {
-      setError("Failed to copy invite links.");
-    }
-  }
-
-  async function copyInviteLink(link: string) {
-    setNoticeArea("student");
-    try {
-      await navigator.clipboard.writeText(link);
-      setSuccess("Invite link copied.");
-    } catch {
-      setError("Failed to copy invite link.");
-    }
-  }
-
-  function escapeCsvCell(value: string): string {
-    const escaped = value.replaceAll('"', '""');
-    return `"${escaped}"`;
-  }
-
-  function downloadCsv(filename: string, header: string[], rows: string[][]) {
-    const lines = [header, ...rows]
-      .map((cols) => cols.map((col) => escapeCsvCell(col)).join(","))
-      .join("\n");
-    const blob = new Blob([lines], { type: "text/csv;charset=utf-8;" });
-    const url = URL.createObjectURL(blob);
-    const anchor = document.createElement("a");
-    anchor.href = url;
-    anchor.download = filename;
-    document.body.append(anchor);
-    anchor.click();
-    anchor.remove();
-    URL.revokeObjectURL(url);
-  }
-
-  const roleLabels: Record<string, string> = {
-    owner: "Owner",
-    co_lecturer: "Co-Lecturer",
-    ta: "TA",
-    student: "Student",
-  };
-
-  const importIssues = useMemo(() => {
-    const rawIssues = (lastImportResult?.issues ?? []) as ImportCsvIssue[];
-    return rawIssues.map((issue, idx) => {
-      const reasonCode = issue.reason || "unknown";
-      const fullName = (issue.full_name ?? "").trim();
-      const studentNumber = (issue.student_number ?? "").trim();
-      const programme = (issue.programme ?? "").trim();
-      const rowNumber = typeof issue.row_number === "number" ? issue.row_number : null;
-      return {
-        key: `${String(issue.email ?? "unknown")}-${reasonCode}-${idx}`,
-        email: issue.email?.trim() ? issue.email : "(unknown)",
-        rowNumber,
-        fullName,
-        studentNumber,
-        programme,
-        reasonCode,
-        reasonLabel: IMPORT_ISSUE_REASON_LABELS[reasonCode] ?? reasonCode,
-      };
-    });
-  }, [lastImportResult?.issues]);
-
-  const previewIssues = useMemo(() => {
-    if (!csvPreview) return [];
-    return csvPreview.issues.map((issue, idx) => ({
-      key: `preview-${issue.rowNumber ?? "header"}-${issue.reasonCode}-${idx}`,
-      email: issue.email?.trim() ? issue.email : "(unknown)",
-      rowNumber: issue.rowNumber,
-      fullName: issue.fullName,
-      studentNumber: issue.studentNumber,
-      programme: issue.programme,
-      reasonCode: issue.reasonCode,
-      reasonLabel: issue.reasonLabel,
-    }));
-  }, [csvPreview]);
-
-  const rosterIssues = importIssues.length > 0 ? importIssues : previewIssues;
-  const issueSourceLabel = importIssues.length > 0 ? "Import issues" : "Preview issues";
-
-  function downloadIssueReportCsv() {
-    if (rosterIssues.length === 0) return;
-    const rows = rosterIssues.map((issue) => [
-      issue.rowNumber === null ? "" : String(issue.rowNumber),
-      issue.email,
-      issue.fullName,
-      issue.studentNumber,
-      issue.programme,
-      issue.reasonCode,
-      issue.reasonLabel,
-    ]);
-    downloadCsv(
-      `roster-import-issues-${course.code.toLowerCase()}.csv`,
-      ["row_number", "email", "name", "student_number", "programme", "reason_code", "reason"],
-      rows
-    );
-    setNoticeArea("student");
-    setSuccess("Issue report CSV downloaded.");
-  }
-
-  function downloadCorrectionCsv() {
-    if (rosterIssues.length === 0) return;
-    const rows = rosterIssues.map((issue) => [
-      issue.email === "(unknown)" ? "" : issue.email,
-      issue.fullName,
-      issue.studentNumber,
-      issue.programme,
-    ]);
-    downloadCsv(
-      `roster-corrections-${course.code.toLowerCase()}.csv`,
-      ["email", "name", "student_number", "programme"],
-      rows
-    );
-    setNoticeArea("student");
-    setSuccess("Correction CSV downloaded. Fix rows and re-import.");
-  }
-
-  return (
-    <div className="flex flex-col gap-8">
-      {/* Notifications */}
-      <div className="order-0 bg-[var(--card)] border border-[var(--border)] rounded-2xl p-5">
-        <div className="flex items-start justify-between gap-4">
-          <div>
-            <h3 className="font-medium text-[var(--foreground)] mb-1">Notifications</h3>
-            <p className="text-xs text-[var(--muted-foreground)]">
-              New submissions are grouped into a digest (10-minute window).
-            </p>
-          </div>
-          <button
-            type="button"
-            onClick={toggleNotifyNewSubmissions}
-            disabled={isSavingNotifyNewSubmissions}
-            className={`relative inline-flex h-6 w-11 items-center rounded-full transition-colors ${notifyNewSubmissions ? "bg-[var(--primary)]" : "bg-[var(--border)]"} disabled:opacity-60`}
-            aria-label="Toggle new submission notifications"
-          >
-            <span
-              className={`inline-block h-4 w-4 transform rounded-full bg-white transition-transform ${notifyNewSubmissions ? "translate-x-6" : "translate-x-1"}`}
-            />
-          </button>
-        </div>
-        {notifyPrefError && (
-          <div className="mt-3 p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-xl text-sm text-[var(--secondary)]">
-            {notifyPrefError}
-          </div>
-        )}
-      </div>
-
-      <div className="order-3 bg-[var(--card)] border border-[var(--border)] rounded-2xl overflow-hidden">
-        <button
-          type="button"
-          onClick={() => setShowStaffSection((v) => !v)}
-          className="w-full p-4 bg-[var(--background)] flex items-center justify-between"
-        >
-          <h3 className="font-medium text-[var(--foreground)]">
-            Staff ({staffMembers.length})
-          </h3>
-          <span className="inline-flex items-center gap-2 text-xs text-[var(--muted-foreground)]">
-            {showStaffSection ? "Hide" : "Show"}
-            {showStaffSection ? (
-              <ChevronUp className="w-4 h-4" />
-            ) : (
-              <ChevronDown className="w-4 h-4" />
-            )}
-          </span>
-        </button>
-      </div>
-
-      {showStaffSection && (
-        <>
-          {/* Add Staff */}
-          <div className="order-4 bg-[var(--card)] border border-[var(--border)] rounded-2xl p-5">
-        <h3 className="font-medium text-[var(--foreground)] mb-1">Add staff</h3>
-        <p className="text-xs text-[var(--muted-foreground)] mb-4">
-          Staff are selected from organization members. Students are enrolled via invites (above).
-        </p>
-        {noticeArea === "staff" && error && (
-          <div className="mb-4 p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-xl text-sm text-[var(--secondary)]">
-            {error}
-          </div>
-        )}
-        {noticeArea === "staff" && success && (
-          <div className="mb-4 p-3 bg-[var(--success)]/10 border border-[var(--success)]/20 rounded-xl text-sm text-[var(--success)]">
-            {success}
-          </div>
-        )}
-        <div className="grid md:grid-cols-12 gap-3 items-end">
-          <div className="md:col-span-5">
-            <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-              Select staff member
-            </label>
-            <select
-              value={selectedUserId ?? ""}
-              onChange={(e) => setSelectedUserId(e.target.value ? Number(e.target.value) : null)}
-              required
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-              disabled={isLoadingOrg}
-            >
-              <option value="">
-                {isLoadingOrg
-                  ? "Loading users..."
-                  : availableOrgMembers.length === 0
-                    ? "No eligible users found"
-                    : "Select a user..."}
-              </option>
-              {availableOrgMembers.map((m) => (
-                <option key={m.id} value={m.user_id}>
-                  {(m.user_email ?? `User #${m.user_id}`)}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className="md:col-span-3">
-            <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-              Staff role
-            </label>
-            <select
-              value={newRole}
-              onChange={(e) => {
-                const value = e.target.value;
-                if (value === "ta" || value === "co_lecturer") setNewRole(value);
-              }}
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-            >
-              <option value="ta">TA</option>
-              <option value="co_lecturer">Co-Lecturer</option>
-            </select>
-          </div>
-          <div className="md:col-span-2">
-            <button
-              onClick={addStaffMember}
-              disabled={isAddingStaff || selectedUserId === null}
-              className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary)]/90 disabled:opacity-60 disabled:cursor-not-allowed transition-colors"
-            >
-              {isAddingStaff ? <Loader2 className="w-4 h-4 animate-spin" /> : <Plus className="w-4 h-4" />}
-              Add staff
-            </button>
-          </div>
-        </div>
-      </div>
-
-      {/* Staff List */}
-      <div className="order-5 bg-[var(--card)] border border-[var(--border)] rounded-2xl overflow-hidden">
-        <div className="p-4 bg-[var(--background)] border-b border-[var(--border)] flex justify-between items-center">
-          <h3 className="font-medium text-[var(--foreground)]">
-            Staff ({staffMembers.length})
-          </h3>
-        </div>
-        <div className="divide-y divide-[var(--border)]">
-          {staffMembers.length === 0 ? (
-            <div className="p-8 text-center text-[var(--muted-foreground)]">No staff members</div>
-          ) : (
-            staffMembers.map((member) => (
-              <div
-                key={member.id}
-                className="flex items-center justify-between p-4 hover:bg-[var(--background)] transition-colors group"
-              >
-                <div className="flex items-center gap-4">
-                  <div className="w-10 h-10 rounded-full bg-[var(--primary)]/10 flex items-center justify-center">
-                    <Users className="w-5 h-5 text-[var(--primary)]" />
-                  </div>
-                  <div>
-                    <p className="font-medium text-[var(--foreground)]">
-                      {member.user_email || `User #${member.user_id}`}
-                    </p>
-                    <p className="text-sm text-[var(--muted-foreground)]">
-                      {roleLabels[member.role]}
-                    </p>
-                  </div>
-                </div>
-                <div className="flex items-center gap-2 opacity-0 group-hover:opacity-100 transition-opacity">
-                  <select
-                    value={member.role}
-                    onChange={(e) => {
-                      const value = e.target.value;
-                      if (value === "owner" || value === "co_lecturer" || value === "ta") {
-                        updateRole(member.id, value);
-                      }
-                    }}
-                    className="text-sm border border-[var(--border)] rounded-lg bg-[var(--background)] px-2 py-1"
-                  >
-                    <option value="owner">Owner</option>
-                    <option value="co_lecturer">Co-Lecturer</option>
-                    <option value="ta">TA</option>
-                  </select>
-                  <button
-                    onClick={() => requestRemoveMember(member.id)}
-                    className="p-2 text-[var(--secondary)] hover:bg-[var(--secondary)]/10 rounded-lg transition-colors"
-                    title="Remove member"
-                  >
-                    <Trash2 className="w-4 h-4" />
-                  </button>
-                </div>
-              </div>
-            ))
-          )}
-        </div>
-      </div>
-
-        </>
-      )}
-
-      {/* Enroll Students */}
-      <div
-        ref={enrollStudentsRef}
-        className="order-1 bg-[var(--card)] border border-[var(--border)] rounded-2xl p-5"
-      >
-        <div className="mb-4">
-          <h3 className="font-medium text-[var(--foreground)] mb-1">Enroll students</h3>
-          <p className="text-xs text-[var(--muted-foreground)]">
-            Use bulk CSV intake for class reps, then use manual invite for one-off corrections.
-          </p>
-        </div>
-
-        {noticeArea === "student" && error && (
-          <div className="mb-4 p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-xl text-sm text-[var(--secondary)]">
-            {error}
-          </div>
-        )}
-        {noticeArea === "student" && success && (
-          <div className="mb-4 p-3 bg-[var(--success)]/10 border border-[var(--success)]/20 rounded-xl text-sm text-[var(--success)]">
-            {success}
-          </div>
-        )}
-
-        <div className="grid xl:grid-cols-5 gap-4">
-          <div className="xl:col-span-2 rounded-2xl border border-[var(--border)] bg-[var(--background)] p-4">
-            <div className="inline-flex items-center gap-2 rounded-full border border-[var(--border)] bg-[var(--card)] px-3 py-1 text-[11px] font-semibold text-[var(--muted-foreground)] uppercase tracking-wider">
-              <Upload className="w-3.5 h-3.5" />
-              Bulk import
-            </div>
-            <p className="mt-3 text-sm text-[var(--foreground)]">
-              Upload roster CSV with profile data for each student.
-            </p>
-            <div className="mt-3 flex flex-wrap gap-2">
-              {["email", "name", "student_number", "programme"].map((column) => (
-                <span
-                  key={column}
-                  className="inline-flex items-center rounded-full border border-[var(--border)] bg-[var(--card)] px-2.5 py-1 text-[11px] font-mono text-[var(--foreground)]"
-                >
-                  {column}
-                </span>
-              ))}
-            </div>
-            <input
-              type="file"
-              ref={fileInputRef}
-              className="hidden"
-              accept=".csv"
-              onChange={handleCsvFileSelected}
-            />
-            <div className="mt-4 grid gap-2">
-              <button
-                ref={importCsvButtonRef}
-                type="button"
-                onClick={() => fileInputRef.current?.click()}
-                disabled={isAddingStudent || isBuildingCsvPreview}
-                className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl border border-[var(--border)] bg-[var(--card)] hover:bg-white disabled:opacity-60 transition-colors text-sm"
-              >
-                {isBuildingCsvPreview ? <Loader2 className="w-4 h-4 animate-spin" /> : <Upload className="w-4 h-4" />}
-                Select CSV
-              </button>
-              <button
-                type="button"
-                onClick={importSelectedCsv}
-                disabled={
-                  isAddingStudent ||
-                  !selectedCsvFile ||
-                  !csvPreview ||
-                  csvPreview.missingHeaders.length > 0
-                }
-                className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary-hover)] disabled:opacity-60 disabled:cursor-not-allowed transition-colors text-sm"
-              >
-                {isAddingStudent ? <Loader2 className="w-4 h-4 animate-spin" /> : <Upload className="w-4 h-4" />}
-                Import previewed CSV
-              </button>
-              <a
-                href="/templates/roster-template.csv"
-                download
-                className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl border border-[var(--border)] bg-[var(--card)] hover:bg-white transition-colors text-sm"
-              >
-                <Download className="w-4 h-4" />
-                Download template
-              </a>
-            </div>
-            <p className="mt-3 text-[11px] text-[var(--muted-foreground)]">
-              Required headers: <code>email,name,student_number,programme</code>
-            </p>
-            {csvPreview && (
-              <div className="mt-3 rounded-xl border border-[var(--border)] bg-[var(--card)] p-3">
-                <p className="text-xs font-semibold text-[var(--foreground)] break-all">
-                  Preview file: {csvPreview.fileName}
-                </p>
-                <p className="mt-1 text-[11px] text-[var(--muted-foreground)]">
-                  Header check:{" "}
-                  {csvPreview.missingHeaders.length === 0 ? (
-                    <span className="text-[var(--success)] font-semibold">Pass</span>
-                  ) : (
-                    <span className="text-[var(--secondary)] font-semibold">
-                      Missing {csvPreview.missingHeaders.join(", ")}
-                    </span>
-                  )}
-                </p>
-                <p className="mt-1 text-[11px] text-[var(--muted-foreground)]">
-                  Rows parsed: {csvPreview.rows.length} | Row issues:{" "}
-                  {csvPreview.issues.filter((issue) => issue.rowNumber !== null).length}
-                </p>
-              </div>
-            )}
-          </div>
-
-          <div className="xl:col-span-3 rounded-2xl border border-[var(--border)] bg-[var(--background)] p-4">
-            <div className="flex items-center justify-between gap-2 mb-3">
-              <p className="text-sm font-semibold text-[var(--foreground)]">Manual invite</p>
-              <p className="required-hint">* Required fields</p>
-            </div>
-            <div className="grid md:grid-cols-12 gap-3 items-end">
-              <div className="md:col-span-4">
-                <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-                  Email
-                </label>
-                <input
-                  value={studentEmail}
-                  onChange={(e) => setStudentEmail(e.target.value)}
-                  placeholder="student@example.com"
-                  required
-                  className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)]"
-                />
-              </div>
-              <div className="md:col-span-4">
-                <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-                  Full name
-                </label>
-                <input
-                  value={studentName}
-                  onChange={(e) => setStudentName(e.target.value)}
-                  placeholder="Jane Doe"
-                  required
-                  className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)]"
-                />
-              </div>
-              <div className="md:col-span-2">
-                <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-                  Student #
-                </label>
-                <input
-                  value={studentNumber}
-                  onChange={(e) => setStudentNumber(e.target.value)}
-                  placeholder="2100714449"
-                  required
-                  className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)]"
-                />
-              </div>
-              <div className="md:col-span-2">
-                <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-                  Programme
-                </label>
-                <select
-                  value={studentProgramme}
-                  onChange={(e) => {
-                    const value = e.target.value;
-                    setStudentProgramme(value ? (value as Programme) : "");
-                  }}
-                  required
-                  className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)]"
-                >
-                  <option value="">Select programme...</option>
-                  {PROGRAMMES.map((p) => (
-                    <option key={p} value={p}>
-                      {p}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              <div className="md:col-span-12">
-                <button
-                  type="button"
-                  onClick={addStudentByEmail}
-                  disabled={
-                    isAddingStudent ||
-                    !studentEmail.trim() ||
-                    !studentName.trim() ||
-                    !studentNumber.trim() ||
-                    !studentProgramme
-                  }
-                  className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary-hover)] disabled:opacity-60 disabled:cursor-not-allowed transition-colors"
-                >
-                  {isAddingStudent ? <Loader2 className="w-4 h-4 animate-spin" /> : <Plus className="w-4 h-4" />}
-                  Enroll / invite student
-                </button>
-                <p className="mt-2 text-[11px] text-[var(--muted-foreground)]">
-                  Existing accounts auto-enroll immediately. New students get invite links valid for 7 days.
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {csvPreview && (
-          <div className="mt-4 rounded-2xl border border-[var(--border)] bg-[var(--background)] p-4">
-            <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2 mb-3">
-              <p className="text-xs uppercase tracking-wider text-[var(--muted-foreground)]">
-                CSV Preview
-              </p>
-              <p className="text-[11px] text-[var(--muted-foreground)]">
-                Standards:{" "}
-                {csvPreview.missingHeaders.length === 0 ? (
-                  <span className="font-semibold text-[var(--success)]">Header check passed</span>
-                ) : (
-                  <span className="font-semibold text-[var(--secondary)]">Header check failed</span>
-                )}
-              </p>
-            </div>
-
-            {csvPreview.rows.length === 0 ? (
-              <p className="text-sm text-[var(--muted-foreground)]">
-                No data rows found in this file.
-              </p>
-            ) : (
-              <div className="overflow-x-auto rounded-xl border border-[var(--border)] bg-[var(--card)]">
-                <table className="w-full min-w-[680px] text-xs">
-                  <thead className="bg-[var(--background)] border-b border-[var(--border)]">
-                    <tr>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Row</th>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Email</th>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Name</th>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Student #</th>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Programme</th>
-                      <th className="text-left px-3 py-2 font-semibold text-[var(--muted-foreground)]">Status</th>
-                    </tr>
-                  </thead>
-                  <tbody className="divide-y divide-[var(--border)]">
-                    {csvPreview.rows.slice(0, 12).map((row) => (
-                      <tr key={`${row.rowNumber}-${row.email}-${row.studentNumber}`}>
-                        <td className="px-3 py-2 text-[var(--muted-foreground)]">{row.rowNumber}</td>
-                        <td className="px-3 py-2 text-[var(--foreground)] break-all">{row.email || "-"}</td>
-                        <td className="px-3 py-2 text-[var(--foreground)]">{row.fullName || "-"}</td>
-                        <td className="px-3 py-2 text-[var(--foreground)]">{row.studentNumber || "-"}</td>
-                        <td className="px-3 py-2 text-[var(--foreground)]">{row.programme || "-"}</td>
-                        <td className="px-3 py-2">
-                          {row.isValid ? (
-                            <span className="inline-flex rounded-full bg-[var(--success)]/15 px-2 py-0.5 text-[10px] font-semibold text-[var(--success)]">
-                              Valid
-                            </span>
-                          ) : (
-                            <span className="inline-flex rounded-full bg-[var(--secondary)]/15 px-2 py-0.5 text-[10px] font-semibold text-[var(--secondary)]">
-                              Needs fix
-                            </span>
-                          )}
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            )}
-
-            {csvPreview.rows.length > 12 && (
-              <p className="mt-2 text-[11px] text-[var(--muted-foreground)]">
-                Showing first 12 rows of {csvPreview.rows.length}.
-              </p>
-            )}
-          </div>
-        )}
-
-        {lastImportResult && (
-          <div className="mt-4 rounded-2xl border border-[var(--border)] bg-[var(--background)] p-4">
-            <p className="text-xs uppercase tracking-wider text-[var(--muted-foreground)] mb-3">
-              Import result
-            </p>
-            <div className="grid sm:grid-cols-3 gap-3">
-              <div className="rounded-xl border border-[var(--border)] bg-[var(--card)] px-3 py-2">
-                <p className="text-[11px] uppercase tracking-wider text-[var(--muted-foreground)]">Invites</p>
-                <p className="text-lg font-semibold text-[var(--foreground)]">{lastImportResult.created_invites}</p>
-              </div>
-              <div className="rounded-xl border border-[var(--border)] bg-[var(--card)] px-3 py-2">
-                <p className="text-[11px] uppercase tracking-wider text-[var(--muted-foreground)]">Auto-enrolled</p>
-                <p className="text-lg font-semibold text-[var(--foreground)]">{lastImportResult.auto_enrolled}</p>
-              </div>
-              <div className="rounded-xl border border-[var(--border)] bg-[var(--card)] px-3 py-2">
-                <p className="text-[11px] uppercase tracking-wider text-[var(--muted-foreground)]">Rows with issues</p>
-                <p className="text-lg font-semibold text-[var(--secondary)]">{importIssues.length}</p>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {rosterIssues.length > 0 && (
-          <div className="mt-4 rounded-2xl border border-[var(--secondary)]/30 bg-[var(--secondary)]/5 p-4">
-            <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2 mb-2">
-              <p className="text-sm font-semibold text-[var(--foreground)]">
-                Rows needing correction ({issueSourceLabel})
-              </p>
-              <div className="flex flex-wrap gap-2">
-                <button
-                  type="button"
-                  onClick={downloadIssueReportCsv}
-                  className="inline-flex items-center gap-1.5 rounded-lg border border-[var(--border)] bg-[var(--card)] px-2.5 py-1.5 text-[11px] text-[var(--foreground)] hover:bg-white transition-colors"
-                >
-                  <Download className="w-3.5 h-3.5" />
-                  Download issues
-                </button>
-                <button
-                  type="button"
-                  onClick={downloadCorrectionCsv}
-                  className="inline-flex items-center gap-1.5 rounded-lg border border-[var(--border)] bg-[var(--card)] px-2.5 py-1.5 text-[11px] text-[var(--foreground)] hover:bg-white transition-colors"
-                >
-                  <Download className="w-3.5 h-3.5" />
-                  Download correction CSV
-                </button>
-              </div>
-            </div>
-            <div className="space-y-2">
-              {rosterIssues.slice(0, 8).map((issue) => (
-                <div
-                  key={issue.key}
-                  className="rounded-lg border border-[var(--secondary)]/20 bg-[var(--card)] px-3 py-2"
-                >
-                  <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-1">
-                    <div className="flex flex-wrap items-center gap-2">
-                      {issue.rowNumber !== null && (
-                        <span className="inline-flex rounded-full bg-[var(--secondary)]/15 px-2 py-0.5 text-[10px] font-semibold text-[var(--secondary)]">
-                          Row {issue.rowNumber}
-                        </span>
-                      )}
-                      <code className="text-xs text-[var(--foreground)] break-all">{issue.email}</code>
-                    </div>
-                    <p className="text-xs text-[var(--secondary)]">{issue.reasonLabel}</p>
-                  </div>
-                  {(issue.fullName || issue.studentNumber || issue.programme) && (
-                    <p className="mt-1 text-[11px] text-[var(--muted-foreground)]">
-                      {issue.fullName ? `Name: ${issue.fullName}` : "Name: -"} |{" "}
-                      {issue.studentNumber ? `Student #: ${issue.studentNumber}` : "Student #: -"} |{" "}
-                      {issue.programme ? `Programme: ${issue.programme}` : "Programme: -"}
-                    </p>
-                  )}
-                </div>
-              ))}
-            </div>
-            {importIssues.length === 0 && (
-              <p className="mt-2 text-[11px] text-[var(--muted-foreground)]">
-                These are preview checks before import. Fix rows and then import.
-              </p>
-            )}
-            {importIssues.length > 0 && (
-              <p className="mt-2 text-[11px] text-[var(--muted-foreground)]">
-                Download correction CSV, fix rows, then import the corrected file.
-              </p>
-            )}
-          </div>
-        )}
-
-        {resolvedInviteLinks.length > 0 && (
-          <div className="mt-4 p-4 bg-[var(--background)] border border-[var(--border)] rounded-2xl">
-            <div className="flex items-start justify-between gap-3 mb-3">
-              <div>
-                <p className="text-sm font-medium text-[var(--foreground)]">Invite links</p>
-                <p className="text-[11px] text-[var(--muted-foreground)]">
-                  No email provider configured yet. Share these links manually with students.
-                </p>
-              </div>
-              <button
-                type="button"
-                onClick={copyInviteLinksAll}
-                className="inline-flex items-center gap-2 px-3 py-2 rounded-xl border border-[var(--border)] bg-[var(--card)] hover:bg-[var(--card)]/70 transition-colors text-xs"
-              >
-                Copy all
-              </button>
-            </div>
-            <div className="space-y-2">
-              {resolvedInviteLinks.map((link) => (
-                <div
-                  key={link}
-                  className="flex items-center justify-between gap-3 px-3 py-2 rounded-xl bg-[var(--card)] border border-[var(--border)]"
-                >
-                  <code className="text-xs text-[var(--foreground)] break-all">{link}</code>
-                  <button
-                    type="button"
-                    onClick={() => copyInviteLink(link)}
-                    className="shrink-0 inline-flex items-center gap-2 px-3 py-2 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary)]/90 transition-colors text-xs"
-                  >
-                    Copy
-                  </button>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-      </div>
-
-      {/* GitHub Classroom Sync */} 
-      <div className="order-1 bg-[var(--card)] border border-[var(--border)] rounded-2xl p-5">
-        <div className="flex items-start justify-between gap-3 mb-4">
-          <div>
-            <h3 className="font-medium text-[var(--foreground)] mb-1">GitHub Classroom sync</h3>
-            <p className="text-xs text-[var(--muted-foreground)]">
-              Bind this course to a GitHub Classroom and compare linked student logins with accepted assignments.
-            </p>
-          </div>
-          <button
-            type="button"
-            onClick={() => void refreshGitHubClassrooms()}
-            disabled={isLoadingClassrooms}
-            className="inline-flex items-center gap-2 px-3 py-2 rounded-xl border border-[var(--border)] bg-[var(--background)] hover:bg-[var(--card)] disabled:opacity-60 transition-colors text-xs"
-          >
-            {isLoadingClassrooms ? <Loader2 className="w-4 h-4 animate-spin" /> : <RefreshCw className="w-4 h-4" />}
-            Refresh classrooms
-          </button>
-        </div>
-
-        {githubClassroomsError && (
-          <div className="mb-4 p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-xl text-sm text-[var(--secondary)]">
-            {githubClassroomsError}
-          </div>
-        )}
-
-        <div className="grid md:grid-cols-12 gap-3 items-end">
-          <div className="md:col-span-8">
-            <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-              Classroom
-            </label>
-            <select
-              value={selectedClassroomId ?? ""}
-              onChange={(e) => setSelectedClassroomId(e.target.value ? Number(e.target.value) : null)}
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)]"
-            >
-              <option value="">Not bound</option>
-              {classrooms.map((classroom) => (
-                <option key={classroom.id} value={classroom.id}>
-                  {classroom.name}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className="md:col-span-4">
-            <button
-              type="button"
-              onClick={() => void saveGitHubClassroomBinding()}
-              disabled={isSavingClassroomBinding}
-              className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary-hover)] disabled:opacity-60 transition-colors"
-            >
-              {isSavingClassroomBinding ? <Loader2 className="w-4 h-4 animate-spin" /> : <Save className="w-4 h-4" />}
-              Save binding
-            </button>
-          </div>
-        </div>
-
-        <div className="grid md:grid-cols-12 gap-3 items-end mt-4">
-          <div className="md:col-span-8">
-            <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">
-              Assignment (for accepted-student check)
-            </label>
-            <select
-              value={selectedGitHubAssignmentId ?? ""}
-              onChange={(e) => setSelectedGitHubAssignmentId(e.target.value ? Number(e.target.value) : null)}
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)]"
-              disabled={!selectedClassroomId}
-            >
-              <option value="">No assignment selected</option>
-              {githubAssignments.map((assignment) => (
-                <option key={assignment.id} value={assignment.id}>
-                  {assignment.title}
-                </option>
-              ))}
-            </select>
-          </div>
-          <div className="md:col-span-4">
-            <button
-              type="button"
-              onClick={() => void runGitHubRosterSync()}
-              disabled={isRunningRosterSync || !selectedClassroomId}
-              className="w-full inline-flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl border border-[var(--border)] bg-[var(--background)] hover:bg-[var(--card)] disabled:opacity-60 transition-colors"
-            >
-              {isRunningRosterSync ? <Loader2 className="w-4 h-4 animate-spin" /> : <RefreshCw className="w-4 h-4" />}
-              Run sync check
-            </button>
-          </div>
-        </div>
-
-        {rosterSync && (
-          <div className="mt-4 p-4 bg-[var(--background)] border border-[var(--border)] rounded-2xl space-y-3">
-            <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-              <div className="p-3 rounded-xl bg-[var(--card)] border border-[var(--border)]">
-                <p className="text-[10px] uppercase tracking-wide text-[var(--muted-foreground)]">Linked</p>
-                <p className="text-lg font-semibold text-[var(--foreground)]">{rosterSync.linked_students_total}</p>
-              </div>
-              <div className="p-3 rounded-xl bg-[var(--card)] border border-[var(--border)]">
-                <p className="text-[10px] uppercase tracking-wide text-[var(--muted-foreground)]">Accepted</p>
-                <p className="text-lg font-semibold text-[var(--foreground)]">{rosterSync.accepted_students_total}</p>
-              </div>
-              <div className="p-3 rounded-xl bg-[var(--card)] border border-[var(--border)]">
-                <p className="text-[10px] uppercase tracking-wide text-[var(--muted-foreground)]">Missing</p>
-                <p className="text-lg font-semibold text-[var(--secondary)]">{rosterSync.missing_logins.length}</p>
-              </div>
-              <div className="p-3 rounded-xl bg-[var(--card)] border border-[var(--border)]">
-                <p className="text-[10px] uppercase tracking-wide text-[var(--muted-foreground)]">Unlinked</p>
-                <p className="text-lg font-semibold text-[var(--foreground)]">{rosterSync.missing_github_students.length}</p>
-              </div>
-            </div>
-
-            {rosterSync.assignments.length > 0 && (
-              <div className="space-y-2">
-                <p className="text-xs font-medium text-[var(--muted-foreground)]">Assignment invite links</p>
-                <div className="space-y-2">
-                  {rosterSync.assignments.slice(0, 6).map((assignment) => (
-                    <div key={assignment.id} className="flex items-center justify-between gap-3 p-3 rounded-xl bg-[var(--card)] border border-[var(--border)]">
-                      <div className="min-w-0">
-                        <p className="text-sm text-[var(--foreground)] truncate">{assignment.title}</p>
-                        {assignment.deadline && (
-                          <p className="text-[10px] text-[var(--muted-foreground)]">
-                            deadline: {new Date(assignment.deadline).toLocaleString()}
-                          </p>
-                        )}
-                      </div>
-                      {assignment.invite_link ? (
-                        <a
-                          href={assignment.invite_link}
-                          target="_blank"
-                          rel="noreferrer"
-                          className="shrink-0 inline-flex items-center gap-1 px-2.5 py-1.5 text-xs font-medium border border-[var(--border)] rounded-lg hover:bg-[var(--background)]"
-                        >
-                          Open invite
-                        </a>
-                      ) : (
-                        <span className="text-[10px] text-[var(--muted-foreground)]">No invite link</span>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      {/* Student List */}
-      <div className="order-2 bg-[var(--card)] border border-[var(--border)] rounded-2xl overflow-hidden">
-        <div className="p-4 bg-[var(--background)] border-b border-[var(--border)]">
-          <h3 className="font-medium text-[var(--foreground)]">
-            Students ({students.length})
-          </h3>
-        </div>
-        <div className="px-4 py-3 bg-[var(--card)] border-b border-[var(--border)] flex items-center justify-between gap-3">
-          <div className="text-xs text-[var(--muted-foreground)]">
-            GitHub linking requests:{" "}
-            <span className="font-medium text-[var(--foreground)]">{pendingGitHubClaims.length}</span>{" "}
-            pending
-          </div>
-          <button
-            type="button"
-            onClick={() => void refreshGitHubClaims()}
-            disabled={isLoadingGitHubClaims}
-            className="inline-flex items-center gap-2 px-3 py-2 rounded-xl border border-[var(--border)] bg-[var(--background)] hover:bg-[var(--card)] disabled:opacity-60 transition-colors text-xs"
-          >
-            {isLoadingGitHubClaims ? (
-              <Loader2 className="w-4 h-4 animate-spin" />
-            ) : (
-              <RefreshCw className="w-4 h-4" />
-            )}
-            Refresh
-          </button>
-        </div>
-        {gitHubClaimsError && (
-          <div className="mx-4 mt-4 p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-xl text-sm text-[var(--secondary)]">
-            {gitHubClaimsError}
-          </div>
-        )}
-        <div className="divide-y divide-[var(--border)]">
-          {students.length === 0 ? (
-            <div className="p-8 text-center">
-              <Users className="w-10 h-10 text-[var(--muted-foreground)] mx-auto mb-3" />
-              <p className="text-[var(--muted-foreground)]">No students enrolled yet</p>
-            </div>
-          ) : (
-            students.map((member) => (
-              <div
-                key={member.id}
-                className="flex items-center justify-between p-4 hover:bg-[var(--background)] transition-colors group"
-              >
-                <div className="flex items-center gap-4">
-                  <div className="w-10 h-10 rounded-full bg-[var(--background)] border border-[var(--border)] flex items-center justify-center">
-                    <Users className="w-5 h-5 text-[var(--muted-foreground)]" />
-                  </div>
-                  <div>
-                    <p className="font-medium text-[var(--foreground)]">
-                      {member.user_email || `User #${member.user_id}`}
-                    </p>
-                    {member.student_number && (
-                      <p className="text-sm text-[var(--muted-foreground)]">
-                        {member.student_number}
-                      </p>
-                    )}
-                  </div>
-                </div>
-
-                <div className="flex items-center gap-3">
-                  {member.github_login ? (
-                    <span className="inline-flex items-center gap-1 px-2.5 py-1 text-xs font-medium bg-[var(--success)]/10 text-[var(--success)] rounded-full">
-                      <LinkIcon className="w-3.5 h-3.5" />
-                      @{member.github_login}
-                    </span>
-                  ) : pendingClaimByMembershipId.has(member.id) ? (
-                    (() => {
-                      const claim = pendingClaimByMembershipId.get(member.id)!;
-                      const isBusy = claimActionId === claim.id;
-                      return (
-                        <div className="flex items-center gap-2">
-                          <span className="inline-flex items-center gap-1 px-2.5 py-1 text-xs font-medium bg-[var(--warning)]/10 text-[var(--warning)] rounded-full">
-                            <LinkIcon className="w-3.5 h-3.5" />
-                            Request: @{claim.github_login}
-                          </span>
-                          <button
-                            onClick={() => void approveGitHubClaim(claim)}
-                            disabled={isBusy}
-                            className="px-2.5 py-1 text-xs font-medium bg-[var(--primary)] text-white rounded-lg hover:opacity-90 transition-opacity disabled:opacity-50"
-                          >
-                            {isBusy ? "Working..." : "Approve"}
-                          </button>
-                          <button
-                            onClick={() => void rejectGitHubClaim(claim)}
-                            disabled={isBusy}
-                            className="px-2.5 py-1 text-xs font-medium border border-[var(--border)] text-[var(--foreground)] rounded-lg hover:bg-[var(--background)] transition-colors disabled:opacity-50"
-                          >
-                            Reject
-                          </button>
-                        </div>
-                      );
-                    })()
-                  ) : (
-                    <span className="inline-flex items-center gap-1 px-2.5 py-1 text-xs font-medium bg-[var(--background)] text-[var(--muted-foreground)] border border-[var(--border)] rounded-full">
-                      <LinkIcon className="w-3.5 h-3.5" />
-                      Not linked
-                    </span>
-                  )}
-
-                  <div className="flex items-center gap-2 opacity-0 group-hover:opacity-100 transition-opacity">
-                    <button
-                      onClick={() => requestRemoveMember(member.id)}
-                      className="p-2 text-[var(--secondary)] hover:bg-[var(--secondary)]/10 rounded-lg transition-colors"
-                      title="Remove student"
-                    >
-                      <Trash2 className="w-4 h-4" />
-                    </button>
-                  </div>
-                </div>
-              </div>
-            ))
-          )}
-        </div>
-      </div>
-
-      <ConfirmModal
-        isOpen={confirmRemoveMembershipId !== null}
-        onClose={() => setConfirmRemoveMembershipId(null)}
-        onConfirm={() => void removeMemberConfirmed()}
-        title="Remove member?"
-        description="This will remove the member from the course. Continue?"
-        confirmLabel="Remove"
-        confirmVariant="danger"
-        isLoading={isRemovingMember}
-      />
-    </div >
-  );
-}
-
-// Assignments Tab
 interface AssignmentsTabProps {
   course: Course;
   assignments: Assignment[];
@@ -3305,1183 +1672,3 @@ function AssignmentsTab({
 }
 
 // Modules Tab with Resources Management
-interface ModulesTabProps {
-  course: Course;
-  modules: Module[];
-  onRefreshModules: () => Promise<void>;
-  openCreateOnMount?: boolean;
-  onConsumedOpenCreate?: () => void;
-}
-
-function ModulesTab({
-  course,
-  modules,
-  onRefreshModules,
-  openCreateOnMount = false,
-  onConsumedOpenCreate,
-}: ModulesTabProps) {
-  const [expandedModules, setExpandedModules] = useState<Set<number>>(new Set());
-  const [showCreateModal, setShowCreateModal] = useState(false);
-  const [newTitle, setNewTitle] = useState("");
-  const [newPosition, setNewPosition] = useState<number>(1);
-  const [isCreating, setIsCreating] = useState(false);
-  const [createError, setCreateError] = useState("");
-
-  const orderedModules = useMemo(
-    () => [...modules].sort((a, b) => a.position - b.position || a.id - b.id),
-    [modules]
-  );
-
-  const nextPosition = useMemo(() => {
-    return orderedModules.length + 1;
-  }, [orderedModules.length]);
-
-  useEffect(() => {
-    if (!showCreateModal) return;
-    setNewTitle("");
-    setNewPosition(nextPosition);
-    setCreateError("");
-  }, [showCreateModal, nextPosition]);
-
-  useEffect(() => {
-    if (!openCreateOnMount) return;
-    setShowCreateModal(true);
-    onConsumedOpenCreate?.();
-  }, [onConsumedOpenCreate, openCreateOnMount]);
-
-  async function createModule() {
-    const title = newTitle.trim();
-    if (!title) return;
-    setIsCreating(true);
-    setCreateError("");
-    try {
-      await courseStaff.createModule(course.id, { title, position: newPosition });
-      await onRefreshModules();
-      setShowCreateModal(false);
-    } catch (err) {
-      if (err instanceof ApiError) setCreateError(err.detail);
-      else setCreateError("Failed to create module");
-    } finally {
-      setIsCreating(false);
-    }
-  }
- 
-  const toggleModule = (moduleId: number) => {
-    setExpandedModules((prev) => {
-      const next = new Set(prev);
-      if (next.has(moduleId)) {
-        next.delete(moduleId);
-      } else {
-        next.add(moduleId);
-      }
-      return next;
-    });
-  };
-
-  return (
-    <div className="space-y-4">
-      {/* Actions */}
-      <div className="flex items-center justify-between">
-        <p className="text-[var(--muted-foreground)]">{modules.length} modules</p>
-        <button
-          onClick={() => setShowCreateModal(true)}
-          className="flex items-center gap-2 px-4 py-2 bg-[var(--primary)] text-white rounded-lg hover:bg-[var(--primary)]/90 transition-colors"
-        >
-          <Plus className="w-4 h-4" />
-          <span>Add Module</span>
-        </button>
-      </div>
-
-      <AnimatePresence>
-        {showCreateModal && (
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black/40"
-            onClick={() => setShowCreateModal(false)}
-          >
-            <motion.div
-              initial={{ opacity: 0, y: 12, scale: 0.98 }}
-              animate={{ opacity: 1, y: 0, scale: 1 }}
-              exit={{ opacity: 0, y: 12, scale: 0.98 }}
-              transition={{ duration: 0.15 }}
-              className="w-full max-w-md bg-[var(--background)] border border-[var(--border)] rounded-2xl shadow-xl"
-              onClick={(e) => e.stopPropagation()}
-            >
-              <div className="p-5 border-b border-[var(--border)] flex items-center justify-between">
-                <div>
-                  <p className="text-sm font-semibold text-[var(--foreground)]">New module</p>
-                  <p className="text-xs text-[var(--muted-foreground)]">{course.code}</p>
-                </div>
-                <button
-                  onClick={() => setShowCreateModal(false)}
-                  className="p-2 rounded-xl hover:bg-[var(--card)] transition-colors"
-                  aria-label="Close"
-                >
-                  <X className="w-4 h-4" />
-                </button>
-              </div>
-
-              <div className="p-5 space-y-4">
-                {createError && (
-                  <div className="p-3 rounded-xl bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 text-sm text-[var(--secondary)]">
-                    {createError}
-                  </div>
-                )}
-
-                <div>
-                  <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">Title</label>
-                  <input
-                    value={newTitle}
-                    onChange={(e) => setNewTitle(e.target.value)}
-                    placeholder="e.g. Week 1: Basics"
-                    required
-                    className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-                  />
-                </div>
-
-                <div>
-                  <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">Position</label>
-                  <input
-                    type="number"
-                    value={newPosition}
-                    onChange={(e) => setNewPosition(Number(e.target.value))}
-                    required
-                    className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-                  />
-                  <p className="mt-2 text-[10px] text-[var(--muted-foreground)]">Suggested next position: {nextPosition}</p>
-                </div>
-              </div>
-
-              <div className="p-5 border-t border-[var(--border)] flex items-center justify-end gap-2">
-                <button
-                  onClick={() => setShowCreateModal(false)}
-                  className="px-4 py-2 rounded-xl border border-[var(--border)] bg-[var(--background)] hover:bg-[var(--card)] transition-colors text-sm"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={createModule}
-                  disabled={isCreating || !newTitle.trim()}
-                  className="inline-flex items-center gap-2 px-4 py-2 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary-hover)] disabled:opacity-60 disabled:cursor-not-allowed transition-colors text-sm"
-                >
-                  {isCreating ? <Loader2 className="w-4 h-4 animate-spin" /> : <Plus className="w-4 h-4" />}
-                  Create
-                </button>
-              </div>
-            </motion.div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-
-      {/* Modules List */}
-      {modules.length === 0 ? (
-        <div className="bg-[var(--card)] border border-[var(--border)] rounded-2xl p-8 text-center">
-          <FolderOpen className="w-10 h-10 text-[var(--muted-foreground)] mx-auto mb-3" />
-          <p className="text-[var(--muted-foreground)]">No modules yet</p>
-          <p className="text-sm text-[var(--muted-foreground)] mt-1">
-            Create modules to organize your course content
-          </p>
-        </div>
-      ) : (
-        <div className="space-y-3">
-          {orderedModules.map((module, idx) => (
-            <ModuleCard
-              key={module.id}
-              courseId={course.id}
-              module={module}
-              isExpanded={expandedModules.has(module.id)}
-              onToggle={() => toggleModule(module.id)}
-              onChanged={onRefreshModules}
-              canMoveUp={idx > 0}
-              canMoveDown={idx < orderedModules.length - 1}
-              onMoveUp={async () => {
-                if (idx <= 0) return;
-                await courseStaff.updateModule(course.id, module.id, { position: idx });
-                await onRefreshModules();
-              }}
-              onMoveDown={async () => {
-                if (idx >= orderedModules.length - 1) return;
-                await courseStaff.updateModule(course.id, module.id, { position: idx + 2 });
-                await onRefreshModules();
-              }}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Module Card with expandable resources section
-interface ModuleCardProps {
-  courseId: number;
-  module: Module;
-  isExpanded: boolean;
-  onToggle: () => void;
-  onChanged: () => Promise<void>;
-  canMoveUp: boolean;
-  canMoveDown: boolean;
-  onMoveUp: () => Promise<void>;
-  onMoveDown: () => Promise<void>;
-}
-
-function ModuleCard(
-  { courseId, module, isExpanded, onToggle, onChanged, canMoveUp, canMoveDown, onMoveUp, onMoveDown }: ModuleCardProps
-) {
-  const [resources, setResources] = useState<ModuleResource[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [showAddModal, setShowAddModal] = useState<"link" | "file" | null>(null);
-  const [showEditModal, setShowEditModal] = useState(false);
-  const [editTitle, setEditTitle] = useState(module.title);
-  const [editPosition, setEditPosition] = useState<number>(module.position);    
-  const [isSavingModule, setIsSavingModule] = useState(false);
-  const [moduleEditError, setModuleEditError] = useState("");
-  const [isMovingModule, setIsMovingModule] = useState(false);
-  const [movingResourceId, setMovingResourceId] = useState<number | null>(null);
-  const [showDeleteModuleConfirm, setShowDeleteModuleConfirm] = useState(false);
-  const [confirmDeleteResourceId, setConfirmDeleteResourceId] = useState<number | null>(null);
-  const [isDeletingModule, setIsDeletingModule] = useState(false);
-  const [isDeletingResource, setIsDeletingResource] = useState(false);
-  const [previewResource, setPreviewResource] = useState<ModuleResource | null>(null);
-  const [previewBlob, setPreviewBlob] = useState<Blob | null>(null);
-  const [isPreviewLoading, setIsPreviewLoading] = useState(false);
-  const [previewError, setPreviewError] = useState("");
-
-  useEffect(() => {
-    setEditTitle(module.title);
-    setEditPosition(module.position);
-  }, [module.id, module.title, module.position]);
-
-  // Fetch resources when expanded
-  useEffect(() => {
-    if (isExpanded && resources.length === 0) {
-      fetchResources();
-    }
-  }, [isExpanded]);
-
-  async function fetchResources() {
-    setIsLoading(true);
-    setError("");
-    try {
-      const data = await courseStaff.listModuleResources(courseId, module.id);
-      setResources(data.sort((a, b) => a.position - b.position || a.id - b.id));
-    } catch (err) {
-      if (err instanceof ApiError) {
-        setError(err.detail);
-      } else {
-        setError("Failed to load resources");
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  }
-
-  async function saveModule() {
-    const title = editTitle.trim();
-    if (!title) return;
-    setIsSavingModule(true);
-    setModuleEditError("");
-    try {
-      await courseStaff.updateModule(courseId, module.id, { title, position: editPosition });
-      await onChanged();
-      setShowEditModal(false);
-    } catch (err) {
-      if (err instanceof ApiError) setModuleEditError(err.detail);
-      else setModuleEditError("Failed to update module");
-    } finally {
-      setIsSavingModule(false);
-    }
-  }
-
-  function requestDeleteModule() {
-    setShowDeleteModuleConfirm(true);
-  }
-
-  async function deleteModuleConfirmed() {
-    setIsDeletingModule(true);
-    setModuleEditError("");
-    try {
-      await courseStaff.deleteModule(courseId, module.id);
-      await onChanged();
-    } catch (err) {
-      if (err instanceof ApiError) setModuleEditError(err.detail);
-      else setModuleEditError("Failed to delete module");
-    } finally {
-      setIsDeletingModule(false);
-      setShowDeleteModuleConfirm(false);
-    }
-  }
-
-  async function handleTogglePublish(resource: ModuleResource) {
-    try {
-      const updated = await courseStaff.updateModuleResource(
-        courseId,
-        module.id,
-        resource.id,
-        { is_published: !resource.is_published }
-      );
-      setResources((prev) =>
-        prev.map((r) => (r.id === resource.id ? updated : r))
-      );
-    } catch (err) {
-      reportError("Failed to toggle publish", err);
-    }
-  }
-
-  async function moveResource(resourceId: number, direction: "up" | "down") {
-    if (movingResourceId !== null) return;
-    const idx = resources.findIndex((r) => r.id === resourceId);
-    if (idx < 0) return;
-    const swapIdx = direction === "up" ? idx - 1 : idx + 1;
-    if (swapIdx < 0 || swapIdx >= resources.length) return;
-
-    const a = resources[idx];
-    const b = resources[swapIdx];
-
-    setMovingResourceId(resourceId);
-    try {
-      const [updatedA, updatedB] = await Promise.all([
-        courseStaff.updateModuleResource(courseId, module.id, a.id, { position: b.position }),
-        courseStaff.updateModuleResource(courseId, module.id, b.id, { position: a.position }),
-      ]);
-      setResources((prev) =>
-        prev
-          .map((r) => (r.id === updatedA.id ? updatedA : r.id === updatedB.id ? updatedB : r))
-          .sort((x, y) => x.position - y.position)
-      );
-    } catch (err) {
-      reportError("Failed to reorder resource", err);
-    } finally {
-      setMovingResourceId(null);
-    }
-  }
-
-  function requestDeleteResource(resourceId: number) {
-    setConfirmDeleteResourceId(resourceId);
-  }
-
-  async function deleteResourceConfirmed() {
-    if (confirmDeleteResourceId === null) return;
-    setIsDeletingResource(true);
-    try {
-      await courseStaff.deleteModuleResource(courseId, module.id, confirmDeleteResourceId);
-      setResources((prev) => prev.filter((r) => r.id !== confirmDeleteResourceId));
-    } catch (err) {
-      reportError("Failed to delete resource", err);
-    } finally {
-      setIsDeletingResource(false);
-      setConfirmDeleteResourceId(null);
-    }
-  }
-
-  async function handleDownload(resource: ModuleResource) {
-    try {
-      const blob = await courseStaff.downloadModuleResource(
-        courseId,
-        module.id,
-        resource.id
-      );
-      const url = window.URL.createObjectURL(blob);
-      const a = document.createElement("a");
-      a.href = url;
-      a.download = resource.file_name || "download";
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      window.URL.revokeObjectURL(url);
-    } catch (err) {
-      reportError("Failed to download resource", err);
-    }
-  }
-
-  function isPdfResource(resource: ModuleResource) {
-    const name = (resource.file_name || "").toLowerCase();
-    return resource.kind === "file" && (name.endsWith(".pdf") || resource.content_type?.includes("pdf"));
-  }
-
-  function closePreview() {
-    setPreviewResource(null);
-    setPreviewBlob(null);
-    setPreviewError("");
-    setIsPreviewLoading(false);
-  }
-
-  async function openPdfPreview(resource: ModuleResource) {
-    if (!isPdfResource(resource)) return;
-    setPreviewResource(resource);
-    setPreviewError("");
-    setIsPreviewLoading(true);
-    try {
-      const blob = await courseStaff.downloadModuleResource(courseId, module.id, resource.id);
-      const isPdf =
-        blob.type.includes("pdf") ||
-        (resource.file_name || "").toLowerCase().endsWith(".pdf");
-      if (!isPdf) {
-        setPreviewBlob(null);
-        setPreviewError("Selected file is not a PDF.");
-        return;
-      }
-      setPreviewBlob(blob);
-    } catch (err) {
-      setPreviewBlob(null);
-      if (err instanceof ApiError) setPreviewError(err.detail);
-      else setPreviewError("Failed to load PDF preview.");
-    } finally {
-      setIsPreviewLoading(false);
-    }
-  }
-
-  function formatFileSize(bytes: number | null) {
-    if (!bytes) return "";
-    if (bytes < 1024) return `${bytes} B`;
-    if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-    return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  }
-
-  const deleteResourceTitle = useMemo(() => {
-    if (confirmDeleteResourceId === null) return null;
-    return resources.find((r) => r.id === confirmDeleteResourceId)?.title ?? null;
-  }, [confirmDeleteResourceId, resources]);
-
-  return (
-    <>
-    <div className="bg-[var(--card)] border border-[var(--border)] rounded-xl overflow-hidden">
-      {/* Module Header */}
-      <div
-        className="flex items-center gap-4 p-4 cursor-pointer hover:bg-[var(--background)] transition-colors"
-        onClick={onToggle}
-      >
-        <div className="w-10 h-10 rounded-lg bg-[var(--primary)]/10 flex items-center justify-center">
-          <span className="text-sm font-semibold text-[var(--primary)]">
-            {module.position}
-          </span>
-        </div>
-        <div className="flex-1 min-w-0">
-          <p className="font-medium text-[var(--foreground)]">{module.title}</p>
-          {module.description && (
-            <p className="text-sm text-[var(--muted-foreground)] truncate">
-              {module.description}
-            </p>
-          )}
-        </div>
-        <div className="flex items-center gap-2">
-          <button
-            onClick={async (e) => {
-              e.stopPropagation();
-              if (!canMoveUp || isMovingModule) return;
-              setIsMovingModule(true);
-              try {
-                await onMoveUp();
-              } finally {
-                setIsMovingModule(false);
-              }
-            }}
-            disabled={!canMoveUp || isMovingModule}
-            className="p-2 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-            title="Move module up"
-            aria-label="Move module up"
-          >
-            <ArrowUp className="w-4 h-4" />
-          </button>
-          <button
-            onClick={async (e) => {
-              e.stopPropagation();
-              if (!canMoveDown || isMovingModule) return;
-              setIsMovingModule(true);
-              try {
-                await onMoveDown();
-              } finally {
-                setIsMovingModule(false);
-              }
-            }}
-            disabled={!canMoveDown || isMovingModule}
-            className="p-2 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-            title="Move module down"
-            aria-label="Move module down"
-          >
-            <ArrowDown className="w-4 h-4" />
-          </button>
-          <button
-            onClick={(e) => {
-              e.stopPropagation();
-              setModuleEditError("");
-              setShowEditModal(true);
-            }}
-            className="p-2 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors"
-          >
-            <Pencil className="w-4 h-4" />
-          </button>
-          <button
-            onClick={(e) => {
-              e.stopPropagation();
-              requestDeleteModule();
-            }}
-            className="p-2 text-[var(--muted-foreground)] hover:text-[var(--secondary)] hover:bg-[var(--secondary)]/10 rounded-lg transition-colors"
-          >
-            <Trash2 className="w-4 h-4" />
-          </button>
-          {isExpanded ? (
-            <ChevronUp className="w-5 h-5 text-[var(--muted-foreground)]" />
-          ) : (
-            <ChevronDown className="w-5 h-5 text-[var(--muted-foreground)]" />
-          )}
-        </div>
-      </div>
-
-      <AnimatePresence>
-        {showEditModal && (
-          <motion.div
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-            className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black/40"
-            onClick={() => setShowEditModal(false)}
-          >
-            <motion.div
-              initial={{ opacity: 0, y: 12, scale: 0.98 }}
-              animate={{ opacity: 1, y: 0, scale: 1 }}
-              exit={{ opacity: 0, y: 12, scale: 0.98 }}
-              transition={{ duration: 0.15 }}
-              className="w-full max-w-md bg-[var(--background)] border border-[var(--border)] rounded-2xl shadow-xl"
-              onClick={(e) => e.stopPropagation()}
-            >
-              <div className="p-5 border-b border-[var(--border)] flex items-center justify-between">
-                <div>
-                  <p className="text-sm font-semibold text-[var(--foreground)]">Edit module</p>
-                  <p className="text-xs text-[var(--muted-foreground)]">{module.title}</p>
-                </div>
-                <button
-                  onClick={() => setShowEditModal(false)}
-                  className="p-2 rounded-xl hover:bg-[var(--card)] transition-colors"
-                  aria-label="Close"
-                >
-                  <X className="w-4 h-4" />
-                </button>
-              </div>
-
-              <div className="p-5 space-y-4">
-                {moduleEditError && (
-                  <div className="p-3 rounded-xl bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 text-sm text-[var(--secondary)]">
-                    {moduleEditError}
-                  </div>
-                )}
-
-                <div>
-                  <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">Title</label>
-                  <input
-                    value={editTitle}
-                    onChange={(e) => setEditTitle(e.target.value)}
-                    required
-                    className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-                  />
-                </div>
-
-                <div>
-                  <label className="block text-xs font-medium text-[var(--muted-foreground)] mb-2">Position</label>
-                  <input
-                    type="number"
-                    value={editPosition}
-                    onChange={(e) => setEditPosition(Number(e.target.value))}
-                    required
-                    className="w-full px-3 py-2.5 bg-[var(--card)] border border-[var(--border)] rounded-xl text-[var(--foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-                  />
-                </div>
-              </div>
-
-              <div className="p-5 border-t border-[var(--border)] flex items-center justify-end gap-2">
-                <button
-                  onClick={() => setShowEditModal(false)}
-                  className="px-4 py-2 rounded-xl border border-[var(--border)] bg-[var(--background)] hover:bg-[var(--card)] transition-colors text-sm"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={saveModule}
-                  disabled={isSavingModule || !editTitle.trim()}
-                  className="inline-flex items-center gap-2 px-4 py-2 rounded-xl bg-[var(--primary)] text-white hover:bg-[var(--primary-hover)] disabled:opacity-60 disabled:cursor-not-allowed transition-colors text-sm"
-                >
-                  {isSavingModule ? <Loader2 className="w-4 h-4 animate-spin" /> : <Save className="w-4 h-4" />}
-                  Save
-                </button>
-              </div>
-            </motion.div>
-          </motion.div>
-        )}
-      </AnimatePresence>
-
-      {/* Expanded Resources Section */}
-      {isExpanded && (
-        <motion.div
-          initial={{ height: 0, opacity: 0 }}
-          animate={{ height: "auto", opacity: 1 }}
-          exit={{ height: 0, opacity: 0 }}
-          className="border-t border-[var(--border)]"
-        >
-          <div className="p-4 bg-[var(--background)]">
-            <div className="flex items-center justify-between mb-3">
-              <h4 className="text-sm font-medium text-[var(--foreground)]">
-                Resources
-              </h4>
-              <div className="flex items-center gap-2">
-                <button
-                  onClick={() => setShowAddModal("link")}
-                  className="flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium text-[var(--primary)] bg-[var(--primary)]/10 hover:bg-[var(--primary)]/20 rounded-lg transition-colors"
-                >
-                  <LinkIcon className="w-3.5 h-3.5" />
-                  Add Link
-                </button>
-                <button
-                  onClick={() => setShowAddModal("file")}
-                  className="flex items-center gap-1.5 px-3 py-1.5 text-xs font-medium text-[var(--primary)] bg-[var(--primary)]/10 hover:bg-[var(--primary)]/20 rounded-lg transition-colors"
-                >
-                  <Upload className="w-3.5 h-3.5" />
-                  Upload File
-                </button>
-              </div>
-            </div>
-
-            {isLoading && (
-              <div className="flex items-center justify-center py-6">
-                <Loader2 className="w-5 h-5 text-[var(--primary)] animate-spin" />
-              </div>
-            )}
-
-            {error && (
-              <div className="p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-lg text-sm text-[var(--secondary)]">
-                {error}
-              </div>
-            )}
-
-            {!isLoading && !error && resources.length === 0 && (
-              <div className="py-6 text-center">
-                <p className="text-sm text-[var(--muted-foreground)]">
-                  No resources added yet
-                </p>
-              </div>
-            )}
-
-            {!isLoading && !error && resources.length > 0 && (
-              <div className="space-y-2">
-                {resources.map((resource, idx) => (
-                  <div
-                    key={resource.id}
-                    className={`flex items-center gap-3 p-3 rounded-lg border transition-colors ${resource.is_published
-                      ? "bg-[var(--card)] border-[var(--border)]"
-                      : "bg-[var(--muted-foreground)]/5 border-dashed border-[var(--muted-foreground)]/30"
-                      }`}
-                  >
-                    <div className="w-8 h-8 rounded-lg bg-[var(--primary)]/10 flex items-center justify-center flex-shrink-0">
-                      {resource.kind === "link" ? (
-                        <LinkIcon className="w-4 h-4 text-[var(--primary)]" />
-                      ) : (
-                        <FileText className="w-4 h-4 text-[var(--primary)]" />
-                      )}
-                    </div>
-                    <div className="flex-1 min-w-0">
-                      <p className="text-sm font-medium text-[var(--foreground)] truncate">
-                        {resource.title}
-                      </p>
-                      <div className="flex items-center gap-2 mt-0.5">
-                        {resource.kind === "link" ? (
-                          <span className="text-xs text-[var(--muted-foreground)] truncate max-w-[200px]">
-                            {resource.url}
-                          </span>
-                        ) : (
-                          <span className="text-xs text-[var(--muted-foreground)]">
-                            {resource.file_name} {resource.size_bytes && `(${formatFileSize(resource.size_bytes)})`}
-                          </span>
-                        )}
-                        {!resource.is_published && (
-                          <span className="px-1.5 py-0.5 text-[10px] font-medium bg-[var(--warning)]/10 text-[var(--warning)] rounded">
-                            Draft
-                          </span>
-                        )}
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-1">
-                      {/* Reorder */}
-                      <button
-                        onClick={() => moveResource(resource.id, "up")}
-                        disabled={idx === 0 || movingResourceId !== null}
-                        className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                        title="Move up"
-                        aria-label="Move resource up"
-                      >
-                        <ArrowUp className="w-4 h-4" />
-                      </button>
-                      <button
-                        onClick={() => moveResource(resource.id, "down")}
-                        disabled={idx === resources.length - 1 || movingResourceId !== null}
-                        className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                        title="Move down"
-                        aria-label="Move resource down"
-                      >
-                        <ArrowDown className="w-4 h-4" />
-                      </button>
-                      {/* Publish/Unpublish */}
-                      <button
-                        onClick={() => handleTogglePublish(resource)}
-                        className={`p-1.5 rounded-lg transition-colors ${resource.is_published
-                          ? "text-emerald-600 hover:bg-emerald-500/10"
-                          : "text-[var(--muted-foreground)] hover:bg-[var(--background)]"
-                          }`}
-                        title={resource.is_published ? "Unpublish" : "Publish"}
-                      >
-                        {resource.is_published ? (
-                          <Eye className="w-4 h-4" />
-                        ) : (
-                          <EyeOff className="w-4 h-4" />
-                        )}
-                      </button>
-
-                      {/* Open/Download */}
-                      {resource.kind === "link" ? (
-                        <a
-                          href={resource.url || "#"}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors"
-                          title="Open link"
-                        >
-                          <ExternalLink className="w-4 h-4" />
-                        </a>
-                      ) : (
-                        <>
-                          {isPdfResource(resource) ? (
-                            <button
-                              onClick={() => void openPdfPreview(resource)}
-                              className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors"
-                              title="Preview PDF"
-                            >
-                              <FileText className="w-4 h-4" />
-                            </button>
-                          ) : null}
-                          <button
-                            onClick={() => handleDownload(resource)}
-                            className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--foreground)] hover:bg-[var(--background)] rounded-lg transition-colors"
-                            title="Download file"
-                          >
-                            <Download className="w-4 h-4" />
-                          </button>
-                        </>
-                      )}
-
-                      {/* Delete */}
-                      <button
-                        onClick={() => requestDeleteResource(resource.id)}
-                        className="p-1.5 text-[var(--muted-foreground)] hover:text-[var(--secondary)] hover:bg-[var(--secondary)]/10 rounded-lg transition-colors"
-                        title="Delete"
-                      >
-                        <Trash2 className="w-4 h-4" />
-                      </button>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </motion.div>
-      )}
-
-      {/* Add Link Modal */}
-      {showAddModal === "link" && (
-        <AddLinkModal
-          courseId={courseId}
-          moduleId={module.id}
-          onClose={() => setShowAddModal(null)}
-          onSuccess={(resource) => {
-            setResources((prev) => [...prev, resource].sort((a, b) => a.position - b.position || a.id - b.id));
-            setShowAddModal(null);
-          }}
-        />
-      )}
-
-      {/* Add File Modal */}
-      {showAddModal === "file" && (
-        <AddFileModal
-          courseId={courseId}
-          moduleId={module.id}
-          onClose={() => setShowAddModal(null)}
-          onSuccess={(resource) => {
-            setResources((prev) => [...prev, resource].sort((a, b) => a.position - b.position || a.id - b.id));
-            setShowAddModal(null);
-          }}
-        />
-      )}
-    </div>
-
-    <PdfPreviewModal
-      isOpen={previewResource !== null}
-      onClose={closePreview}
-      title={previewResource?.title || "PDF preview"}
-      fileName={previewResource?.file_name}
-      blob={previewBlob}
-      isLoading={isPreviewLoading}
-      error={previewError}
-      onRetry={
-        previewResource
-          ? () => {
-              void openPdfPreview(previewResource);
-            }
-          : undefined
-      }
-      onDownload={
-        previewResource
-          ? () => {
-              void handleDownload(previewResource);
-            }
-          : undefined
-      }
-    />
-
-    <ConfirmModal
-      isOpen={showDeleteModuleConfirm}
-      onClose={() => setShowDeleteModuleConfirm(false)}
-      onConfirm={() => void deleteModuleConfirmed()}
-      title={`Delete \"${module.title}\"?`}
-      description="This removes module content ordering (resources and assignments remain). Continue?"
-      confirmLabel="Delete"
-      confirmVariant="danger"
-      isLoading={isDeletingModule}
-    />
-
-    <ConfirmModal
-      isOpen={confirmDeleteResourceId !== null}
-      onClose={() => setConfirmDeleteResourceId(null)}
-      onConfirm={() => void deleteResourceConfirmed()}
-      title={deleteResourceTitle ? `Delete \"${deleteResourceTitle}\"?` : "Delete resource?"}
-      description="This cannot be undone. Continue?"
-      confirmLabel="Delete"
-      confirmVariant="danger"
-      isLoading={isDeletingResource}
-    />
-    </>
-  );
-}
-
-// Add Link Modal
-interface AddLinkModalProps {
-  courseId: number;
-  moduleId: number;
-  onClose: () => void;
-  onSuccess: (resource: ModuleResource) => void;
-}
-
-function AddLinkModal({ courseId, moduleId, onClose, onSuccess }: AddLinkModalProps) {
-  const [title, setTitle] = useState("");
-  const [url, setUrl] = useState("");
-  const [isPublished, setIsPublished] = useState(true);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState("");
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!title.trim() || !url.trim()) return;
-
-    setIsSubmitting(true);
-    setError("");
-
-    try {
-      const resource = await courseStaff.createLinkResource(courseId, moduleId, {
-        title: title.trim(),
-        url: url.trim(),
-        is_published: isPublished,
-      });
-      onSuccess(resource);
-    } catch (err) {
-      if (err instanceof ApiError) {
-        setError(err.detail);
-      } else {
-        setError("Failed to add link");
-      }
-    } finally {
-      setIsSubmitting(false);
-    }
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-      <div className="fixed inset-0 bg-black/50" onClick={onClose} />
-      <motion.div
-        initial={{ opacity: 0, scale: 0.95 }}
-        animate={{ opacity: 1, scale: 1 }}
-        className="relative w-full max-w-md bg-[var(--card)] border border-[var(--border)] rounded-2xl shadow-xl"
-      >
-        <div className="flex items-center justify-between p-4 border-b border-[var(--border)]">
-          <h3 className="font-[family-name:var(--font-display)] font-semibold text-[var(--foreground)]">
-            Add Link Resource
-          </h3>
-          <button
-            onClick={onClose}
-            className="p-1 text-[var(--muted-foreground)] hover:text-[var(--foreground)] rounded-lg transition-colors"
-          >
-            <X className="w-5 h-5" />
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="p-4 space-y-4">
-          {error && (
-            <div className="p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-lg text-sm text-[var(--secondary)]">
-              {error}
-            </div>
-          )}
-
-          <div>
-            <label className="block text-sm font-medium text-[var(--foreground)] mb-1.5">
-              Title
-            </label>
-            <input
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="e.g., Course Syllabus"
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-              required
-            />
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-[var(--foreground)] mb-1.5">
-              URL
-            </label>
-            <input
-              type="url"
-              value={url}
-              onChange={(e) => setUrl(e.target.value)}
-              placeholder="https://..."
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-              required
-            />
-          </div>
-
-          <label className="flex items-center gap-2 cursor-pointer">
-            <input
-              type="checkbox"
-              checked={isPublished}
-              onChange={(e) => setIsPublished(e.target.checked)}
-              className="w-4 h-4 rounded border-[var(--border)] text-[var(--primary)] focus:ring-[var(--primary)]"
-            />
-            <span className="text-sm text-[var(--foreground)]">
-              Publish immediately (visible to students)
-            </span>
-          </label>
-
-          <div className="flex items-center gap-3 pt-2">
-            <button
-              type="button"
-              onClick={onClose}
-              className="flex-1 px-4 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] hover:bg-[var(--card)] transition-colors"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              disabled={isSubmitting || !title.trim() || !url.trim()}
-              className="flex-1 px-4 py-2.5 bg-[var(--primary)] text-white rounded-xl hover:bg-[var(--primary)]/90 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {isSubmitting ? (
-                <Loader2 className="w-4 h-4 mx-auto animate-spin" />
-              ) : (
-                "Add Link"
-              )}
-            </button>
-          </div>
-        </form>
-      </motion.div>
-    </div>
-  );
-}
-
-// Add File Modal
-interface AddFileModalProps {
-  courseId: number;
-  moduleId: number;
-  onClose: () => void;
-  onSuccess: (resource: ModuleResource) => void;
-}
-
-function AddFileModal({ courseId, moduleId, onClose, onSuccess }: AddFileModalProps) {
-  const [title, setTitle] = useState("");
-  const [file, setFile] = useState<File | null>(null);
-  const [isPublished, setIsPublished] = useState(true);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState("");
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  async function handleSubmit(e: React.FormEvent) {
-    e.preventDefault();
-    if (!title.trim() || !file) return;
-
-    setIsSubmitting(true);
-    setError("");
-
-    try {
-      const resource = await courseStaff.uploadFileResource(
-        courseId,
-        moduleId,
-        file,
-        title.trim(),
-        undefined,
-        isPublished
-      );
-      onSuccess(resource);
-    } catch (err) {
-      if (err instanceof ApiError) {
-        setError(err.detail);
-      } else {
-        setError("Failed to upload file");
-      }
-    } finally {
-      setIsSubmitting(false);
-    }
-  }
-
-  function handleFileSelect(e: React.ChangeEvent<HTMLInputElement>) {
-    const selectedFile = e.target.files?.[0];
-    if (selectedFile) {
-      setFile(selectedFile);
-      if (!title) {
-        // Auto-fill title from filename
-        const name = selectedFile.name.replace(/\.[^/.]+$/, ""); // Remove extension
-        setTitle(name);
-      }
-    }
-  }
-
-  function formatFileSize(bytes: number) {
-    if (bytes < 1024) return `${bytes} B`;
-    if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-    return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-      <div className="fixed inset-0 bg-black/50" onClick={onClose} />
-      <motion.div
-        initial={{ opacity: 0, scale: 0.95 }}
-        animate={{ opacity: 1, scale: 1 }}
-        className="relative w-full max-w-md bg-[var(--card)] border border-[var(--border)] rounded-2xl shadow-xl"
-      >
-        <div className="flex items-center justify-between p-4 border-b border-[var(--border)]">
-          <h3 className="font-[family-name:var(--font-display)] font-semibold text-[var(--foreground)]">
-            Upload File Resource
-          </h3>
-          <button
-            onClick={onClose}
-            className="p-1 text-[var(--muted-foreground)] hover:text-[var(--foreground)] rounded-lg transition-colors"
-          >
-            <X className="w-5 h-5" />
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="p-4 space-y-4">
-          {error && (
-            <div className="p-3 bg-[var(--secondary)]/10 border border-[var(--secondary)]/20 rounded-lg text-sm text-[var(--secondary)]">
-              {error}
-            </div>
-          )}
-
-          {/* File Drop Zone */}
-          <div
-            onClick={() => fileInputRef.current?.click()}
-            className={`relative border-2 border-dashed rounded-xl p-6 text-center cursor-pointer transition-colors ${file
-              ? "border-[var(--primary)] bg-[var(--primary)]/5"
-              : "border-[var(--border)] hover:border-[var(--primary)]/50 hover:bg-[var(--background)]"
-              }`}
-          >
-            <input
-              ref={fileInputRef}
-              type="file"
-              onChange={handleFileSelect}
-              className="hidden"
-            />
-            {file ? (
-              <div className="flex items-center justify-center gap-3">
-                <FileText className="w-8 h-8 text-[var(--primary)]" />
-                <div className="text-left">
-                  <p className="text-sm font-medium text-[var(--foreground)] truncate max-w-[200px]">
-                    {file.name}
-                  </p>
-                  <p className="text-xs text-[var(--muted-foreground)]">
-                    {formatFileSize(file.size)}
-                  </p>
-                </div>
-                <button
-                  type="button"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    setFile(null);
-                  }}
-                  className="p-1 text-[var(--muted-foreground)] hover:text-[var(--secondary)]"
-                >
-                  <X className="w-4 h-4" />
-                </button>
-              </div>
-            ) : (
-              <>
-                <Upload className="w-8 h-8 text-[var(--muted-foreground)] mx-auto mb-2" />
-                <p className="text-sm text-[var(--foreground)]">
-                  Click to select a file
-                </p>
-                <p className="text-xs text-[var(--muted-foreground)] mt-1">
-                  PDF, DOC, TXT, ZIP, etc.
-                </p>
-              </>
-            )}
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-[var(--foreground)] mb-1.5">
-              Title
-            </label>
-            <input
-              type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              placeholder="e.g., Lecture Notes Week 1"
-              className="w-full px-3 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] placeholder:text-[var(--muted-foreground)] focus:outline-none focus:ring-2 focus:ring-[var(--primary)]"
-              required
-            />
-          </div>
-
-          <label className="flex items-center gap-2 cursor-pointer">
-            <input
-              type="checkbox"
-              checked={isPublished}
-              onChange={(e) => setIsPublished(e.target.checked)}
-              className="w-4 h-4 rounded border-[var(--border)] text-[var(--primary)] focus:ring-[var(--primary)]"
-            />
-            <span className="text-sm text-[var(--foreground)]">
-              Publish immediately (visible to students)
-            </span>
-          </label>
-
-          <div className="flex items-center gap-3 pt-2">
-            <button
-              type="button"
-              onClick={onClose}
-              className="flex-1 px-4 py-2.5 bg-[var(--background)] border border-[var(--border)] rounded-xl text-[var(--foreground)] hover:bg-[var(--card)] transition-colors"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              disabled={isSubmitting || !title.trim() || !file}
-              className="flex-1 px-4 py-2.5 bg-[var(--primary)] text-white rounded-xl hover:bg-[var(--primary)]/90 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {isSubmitting ? (
-                <Loader2 className="w-4 h-4 mx-auto animate-spin" />
-              ) : (
-                "Upload File"
-              )}
-            </button>
-          </div>
-        </form>
-      </motion.div>
-    </div>
-  );
-}
-
